+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "bluejay-printer",
     "bluejay-schema-comparator",
     "bluejay-typegen",
+    "bluejay-typegen-codegen",
     "bluejay-typegen-macro",
     "bluejay-validator",
     "bluejay-visibility",
@@ -23,6 +24,7 @@ bluejay-parser = { path = "./bluejay-parser", version = "=0.2.0" }
 bluejay-printer = { path = "./bluejay-printer", version = "=0.2.0" }
 bluejay-schema-comparator = { path = "./bluejay-schema-comparator", version = "=0.2.0" }
 bluejay-typegen = { path = "./bluejay-typegen", version = "=0.2.0" }
+bluejay-typegen-codegen = { path = "./bluejay-typegen-codegen", version = "=0.2.0" }
 bluejay-typegen-macro = { path = "./bluejay-typegen-macro", version = "=0.2.0" }
 bluejay-validator = { path = "./bluejay-validator", version = "=0.2.0" }
 bluejay-visibility = { path = "./bluejay-visibility", version = "=0.2.0" }
