+++ bluejay-typegen-codegen/Cargo.toml
@@ -1,24 +1,23 @@
 [package]
-name = "bluejay-typegen-macro"
+name = "bluejay-typegen-codegen"
 version.workspace = true
 edition = "2021"
 license = "MIT"
 repository = "https://github.com/Shopify/bluejay"
 homepage = "https://github.com/Shopify/bluejay"
-keywords = ["graphql"]
+keywords = ["graphql", "codegen"]
 exclude = [".gitignore", "tests/**/*"]
-description = "Macros for `bluejay-typegen`"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+description = "Code generation utilities for `bluejay-typegen-macro`"
 
 [dependencies]
-bluejay-typegen-codegen = { workspace = true }
+bluejay-core = { workspace = true }
+bluejay-parser = { workspace = true, features = ["format-errors"] }
+bluejay-validator = { workspace = true }
 syn = { version = "2.0", features = ["full"] }
 quote = "1.0"
 proc-macro2 = "1.0"
-
-[lib]
-proc-macro = true
+convert_case = "0.8"
+itertools = "0.14.0"
 
 [features]
 default = ["serde"]
