+++ bluejay-typegen-codegen/src/attributes.rs
