+++ bluejay-typegen-codegen/src/builtin_scalar.rs
