+++ bluejay-typegen-codegen/src/enum_type_definition.rs
