+++ bluejay-typegen-codegen/src/executable_definition/executable_enum_builder.rs
