+++ bluejay-typegen-codegen/src/executable_definition/executable_struct_builder.rs
