+++ bluejay-typegen-codegen/src/names.rs
