+++ bluejay-typegen-codegen/src/validation.rs
