+++ bluejay-typegen-codegen/src/validation/error.rs
