+++ bluejay-typegen-codegen/src/validation/fragment_and_operation_names_do_not_clash.rs
