+++ bluejay-typegen-codegen/src/validation/selections_are_valid.rs
